requirements.txt
django<1.10
psycopg2==2.6
django-extensions==1.6.1
python-social-auth==0.2.13
requests-oauthlib==0.6.0
gunicorn==19.4.1
django-forms-bootstrap==3.0.1
rules==1.1.1
opbeat==3.1.3
pytest==2.8.5
pytest-django==2.8.0
tox==2.3.1
pylama==7.0.7
djangorestframework==3.3.3
Pillow

